Merge in PE_RBVD_APP-ID-105529_DSG/dto_enterprise_life_insurance from feature/updateGetQuotationDTO to develop

* commit '75c0d25c83707686a645a60ca09dd1a83d0ba5ce':
  feat: PERAR-26254 PERAR-26864 New classes for quotation detail rimac api
  feat: PERAR-26254 PERAR-26864 Update parameteres in GetQuotationDTO class

+++ src/main/java/com/bbva/rbvd/dto/enterpriseinsurance/commons/dto/AmountDTO.java
@@ -1,20 +1,20 @@
 package com.bbva.rbvd.dto.enterpriseinsurance.commons.dto;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 public class AmountDTO implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private BigDecimal amount;
+    private Double amount;
     private String currency;
 
-    public BigDecimal getAmount() {
+
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
